+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 
 ### API
+* add English annotations to symbols in vocabulary
+* fix bug with Fresh_Variable
 
 ### GUI